add new file parseSchematicColor.ts
